fix some errors with batterys
